ssntp: Support legacy configuration payloads
We recently changed the configuration payload to support arrays
of compute and management subnets. We may change it in the short
term to support our storage solution.
To give us more flexibility, we want to be able to support legacy
payloads when we are forced to break configuration backward
compatibility.

Fixes #457

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>