Introduce new Ciao Preparation functionality
This change introduces a minor subcomponent, `prepare`, which facilitates the
inclusion of distro/host-specific configurations at runtime. Given that each
component of Ciao may have its own set of dependencies, and that each distro
can have a different way to express these relationships and operations, we
make them highly generic through the PackageRequirements type.

Users of the API populate a PackageRequirements table with distro mappings
of binary path to package name/bundle name, which allows the core detection
system to use the appropriate functionality to find out if the package(s) are
installed, and take action if not. Helper functions are also included for
less generic items, such as the handling of Docker and Qemu minimum version
requirements.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>