osprepare: Ensure its ok to fail early
It is quite likely that we either encounter a distro or configuration that is
new to us, due to the nature of the new osprepare module, so we shouldn't
fail here. Rather, we should fail later on so that osprepare can be improved
retroactively to fix whatever setup issue we encountered.

With this change it is also OK to pass nil dependencies, to enable easier
plugging in of the functionality, and future expansion of the core entry
method, PrepareOsDeps.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>